reduce git corruption issues
On RasperryPI, we seee many instances where the git repo has become corrupt.
In several instances, I found that the reflog simply had a bad line, which once
removed caused things to work, at other times, I found that the tree had
missing blobs.

Should help with https://github.com/openaps/openaps/issues/30
This fixes incorrect usage of the GitPython API, according to
      https://github.com/gitpython-developers/GitPython/issues/265

In our usage, each report is run in sequence.  After each report is run, the
result is added to git index, but the result is not committed until all the
reports have finished.  This means that GitPython is writing out incorrect tree
data until all the reports have finished... if any of the reports fail or cause
the process to stop (due to battery, etc), then the git repo will have a
tainted/corrupt tree, missing blogs, etc.

Incorrect:
  * Using `IndexFile.add` without a matching `IndexFile.commit`
  * mixing `IndexFile.add` with `git commit`

According to the GitPython bug report, git corruption can be expected,
especially with older versions.  The issue was aggravated by mixing usage, and
creating a delay between when a file was added to index and when it was
committed.

Correct:
  * Matching use of `IndexFile.{add,commit}`
  * Matching use of `git` vs `IndexFile`
  * Use `git add` which was designed to work in pair-wise operation with `git
    commit`, this is especially helpful as a workaround on older versions.

CC. @scottleibrand, @amazaheri, @openaps/all