Handle pump session expiry that is far in the future
When the pi/edison boots up, the time may be very far in the past (eg: 2000).

The expiry value is thus considered 'in the future' and the system will not attempt
to wake the pump up. This means that even if the loop.sh script tries to force the
time based on the pump, it will fail if the pump session has expired.

We now override any expiry in the future if it's far in the future, so that we
can correct the time of the Linux system automatically